Refactor mobile nav modal
