+++ life-map/src/components/NavBar/FollowRequestCard.js
